pyproject.toml
[project]
name = "integrated-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.111.0",
    "uvicorn[standard]>=0.29.0",
    "langchain>=0.2.0",
    "langchain-groq>=0.1.5",
    "langchain-core>=0.2.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.7.0",
    "groq>=0.9.0",
    "supabase>=2.5.0",
]
